Set version to snapshot
The version needs to be set to SNAPSHOT for Travis-CI to deploy builds regularly, else you end up with no snapshot builds to use with Maven. Release versioning should only be used for a single commit meant for performing a stable release (not experimental changes or untested features), which is meant to then be deployed to the staging repository at oss.sonatype.org. The last snapshot deployed was in January.